// migration/types/screenplay.ts
export type LineType =
  | "action"
  | "dialogue"
  | "character"
  | "scene-header-1"
  | "scene-header-2"
  | "scene-header-3"
  | "scene-header-top-line"
  | "transition"
  | "parenthetical"
  | "basmala";